requirements.txt
celery==4.2.0
Django==2.0.7
djangorestframework==3.8.2
requests==2.19.1
pyopenssl
